Refactor annotations and validators structure.
Moved `AgeRange` and related classes to `shared.annotations` for better modularization and reused `ValidIbanValidator` to replace `IbanValidator`. Updated corresponding imports and references to ensure consistency across the application.
