+++ src/test/java/info/mackiewicz/bankapp/core/user/validation/AgeRangeValidatorTest.java
@@ -1,5 +1,6 @@
 package info.mackiewicz.bankapp.core.user.validation;
 
+import info.mackiewicz.bankapp.shared.annotations.AgeRangeValidatior;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
